- fix third-party style enqueue - update third-party index - split Explorer into its own components - move Explorer utils into its own file - simplify App.jsx - move Query Composer styles from main plugin into 3rd party
