+++ plugins/third-party-plugin/src/components/ArrowOpen.jsx
@@ -1,0 +1,18 @@
+const ArrowOpen = () => (
+	<svg
+		viewBox="0 -4 13 15"
+		style={{
+			color: 'hsla(var(--color-neutral), var(--alpha-tertiary, 0.4))',
+			marginRight: 'var(--px-4)',
+			height: 'var(--px-16)',
+			width: 'var(--px-16)',
+		}}
+	>
+		<path
+			d="M3.35355 6.85355L6.14645 9.64645C6.34171 9.84171 6.65829 9.84171 6.85355 9.64645L9.64645 6.85355C9.96143 6.53857 9.73835 6 9.29289 6L3.70711 6C3.26165 6 3.03857 6.53857 3.35355 6.85355Z"
+			fill="currentColor"
+		/>
+	</svg>
+);
+
+export default ArrowOpen;
