+++ plugins/third-party-plugin/src/components/DefaultArrowClosed.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function DefaultArrowOpen(){
+	return (
+		<svg width="12" height="9">
+			<path fill="#666" d="M 0 0 L 0 9 L 5.5 4.5 z" />
+		</svg>
+	);
+}
