+++ plugins/third-party-plugin/src/components/DefaultArrowOpen.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function DefaultArrowOpen(){
+	return (
+		<svg width="12" height="9">
+			<path fill="#666" d="M 0 2 L 9 2 L 4.5 7.5 z" />
+		</svg>
+	);
+}
