+++ plugins/third-party-plugin/src/index.js
@@ -1,71 +1,33 @@
 /**
  * External dependencies
  */
-import { explorerPlugin } from '@graphiql/plugin-explorer';
-import { addAction } from '@wordpress/hooks';
 import {QueryComposer} from "./components/QueryComposer";
-
-console.log( 'Third Party Plugin loading...' );
+import {edit as editIcon, Icon} from '@wordpress/icons';
 
 window.addEventListener('WPGraphQLIDE_Window_Ready', function(event) {
-	console.log( 'Third Party Plugin Loaded' );
 
 	if ( ! window.WPGraphQLIDE ) {
-		console.log( `WPGraphQLIDE not loaded` );
-		console.log( window.WPGraphQLIDE );
 		return;
 	}
 
 	const { registerActivityBarPanel } = window.WPGraphQLIDE || {};
 
 	if ( typeof registerActivityBarPanel === 'function' ) {
-
-		console.log( `Registering Mock Explorer Panel`);
-
 		registerActivityBarPanel( 'query-composer', () => {
 			return {
 				title: 'Query Composer',
-				icon: () => <>Query Composer</>,
-				content: () => (
-					<section aria-label="Query Composer" className="graphiql-query-composer">
-						<QueryComposer/>
-					</section>
-				)
+				icon: () => (
+					<Icon
+						icon={ editIcon }
+						style={ {
+							fill: 'hsla(var(--color-neutral), var(--alpha-tertiary))',
+						} }
+					/>
+				),
+				content: () => <QueryComposer />
 			};
 		}, 4 );
-
-		// registerActivityBarPanel( 'explorer', explorerPlugin, 4 );
-	} else {
-		console.log( 'CANNOT Registering Mock Explorer Panel' );
 	}
-
-	// Perform your desired actions here
 });
-
-
-
-
-
-
-
-// return;
-//
-
-// const mockExplorerPanel = () => {
-// 	return {
-// 		title: 'Mock Explorer',
-// 		icon: 'icon',
-// 		content: () => (
-// 			<>
-// 				<h2>Test</h2>
-// 			</>
-// 		),
-// 	};
-// };
-//
-// /**
-//  * Ensure the code only runs once WPGraphQLIDEReady is fired and registerActivityBarPanel is defined.
-//  */
-
