+++ plugins/third-party-plugin/third-party-plugin.php
@@ -30,12 +30,12 @@ add_action( 'wpgraphqlide_enqueue_script', function() {
         $asset_file['version']
     );
 
-    wp_enqueue_style(
-        'third-party-plugin',
-        plugins_url( 'build/index.css', __FILE__ ),
-        [],
-        $asset_file['version'],
-    );
+	wp_enqueue_style(
+		'third-party-plugin',
+		plugins_url( 'build/style-index.css', __FILE__ ),
+		[],
+		$asset_file['version'],
+	);
 
 });
